+++ lib/page/mod.ts
@@ -2,6 +2,7 @@ export * from "./model.ts";
 export * from "./interfaces.ts";
 export * from "./cursor.ts";
 export * from "./endpoint.ts";
+export * from "./endpoint-json.ts";
 export * from "./aggregate.ts";
 export * from "./aggregate-interfaces.ts";
 export * from "./memory-aggregate-repository.ts";